+++ src/main/java/com/vinicius/vs/agenda/enumeration/AgendaResult.java
@@ -1,0 +1,7 @@
+package com.vinicius.vs.agenda.enumeration;
+
+public enum AgendaResult {
+
+    SIM,
+    NAO
+}