+++ src/main/java/com/vinicius/vs/constants/RabbitmqConstantes.java
@@ -1,0 +1,10 @@
+package com.vinicius.vs.constants;
+
+public class RabbitmqConstantes {
+
+    public static final String QEUE_AGENDA = "agenda.v1.result";
+    public static final String EXCHANGE_AGENDA = "agenda.direct";
+
+    public static final String QEUE_ASSOCIATE = "associate.v1.create";
+    public static final String EXCHANGE_ASSOCIATE = "associate.direct";
+}