+++ src/main/java/com/vinicius/vs/vote/dtos/VoteConfirmationDTO.java
@@ -1,0 +1,20 @@
+package com.vinicius.vs.vote.dtos;
+
+import com.vinicius.vs.agenda.enumeration.AgendaResult;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class VoteConfirmationDTO {
+
+    private String name;
+
+    private AgendaResult vote;
+
+    private String subject;
+}