+++ src/main/java/com/vinicius/vs/voto/repository/IVotoRepository.java
@@ -1,9 +1,0 @@
-package com.vinicius.vs.voto.repository;
-
-import com.vinicius.vs.voto.model.Voto;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IVotoRepository extends JpaRepository<Voto, Long> {
-}