+++ src/main/java/com/vinicius/vs/voto/service/IVotoService.java
@@ -1,8 +1,0 @@
-package com.vinicius.vs.voto.service;
-
-import com.vinicius.vs.voto.dtos.ConfirmacaoVoto;
-import com.vinicius.vs.voto.dtos.VotarDTO;
-
-public interface IVotoService {
-    ConfirmacaoVoto votar(VotarDTO votarDTO);
-}